Quest System Fisrt Integration
I've done some simple Quest System
assumptions. I've tried to add some sample quest, but there is a problem with logical conditions like and which are not doing the job properly. I will continue to solve this problem in this week, and I will add first sample quest by the next week.